backend/requirements.txt
Flask==3.0.3
gunicorn==21.2.0
celery==5.3.6
redis==5.0.1
psycopg2-binary==2.9.11
SQLAlchemy==2.0.44
python-dotenv==1.1.1
boto3==1.34.131
PyJWT==2.8.0
bcrypt==4.1.2
flask-cors==4.0.0
werkzeug==3.1.3
